Internal Documentation

Internal documentation for SymplecticMapTools.jl

Contents

Internal Documentation

Contents

Index

Fourier Circles

Birkhoff Extrapolation

Index

SymplecticMapTools.cheb_roots_to_roots

SymplecticMapTools.chebyshev_companion_matrix

SymplecticMapTools.evaluate

SymplecticMapTools.get_N

SymplecticMapTools.palindromic_cheb_roots

SymplecticMapTools.palindromic_to_chebyshev

Fourier Circles

SymplecticMapTools.get_N — Function

get_N(z::FourierCircle)

Get number of unknown parameters per circle in island, excluding τ

SymplecticMapTools.evaluate — Function

evaluate(z::FourierCircle, θ::AbstractVector{T}; i_circle::Integer=1) where {T}

Evaluate the i_circleth circle in the chain at a vector of points θ

evaluate(z::InvariantCircle, θ::Number; i_circle::Integer=1)

Evaluate the i_circleth circle in z at the point θ

evaluate(c::ConnectingOrbit, x::AbstractArray; i_p::Integer=1)

Evaluate the i_pth connecting orbit at a set of points x[j] in [0,1]

evaluate(c::ConnectingOrbit, x::Number; i_p::Integer=1)

Evaluate the i_pth connecting orbit at a point x in [0,1]

evaluate(k::KernelLabel, x::AbstractArray)

Evaluate the kernel matrix at the columns of x

evaluate(c::ContFrac)

Find a floating point representation of the continued fraction.

Birkhoff Extrapolation

SymplecticMapTools.palindromic_to_chebyshev — Function

palindromic_to_chebyshev(c::AbstractVector)

Input:

c: A palindromic set of monomial coefficients for polynomial p(z)

Output:

v: The coefficients of a Chebyshev polynomial q s.t. q((z+inv(z))/2) = p(z)

SymplecticMapTools.palindromic_cheb_roots — Function

palindromic_cheb_roots(c::AbstractVector)

Get the roots of the chebyshev polynomial associated with the palindromic polynomial with coefficients c.

Input:

c: A palindromic set of monomial coefficients for polynomial p(z)

Output:

v: The roots of the polynomial q satisfying q((z+inv(z))/2) = p(z)

SymplecticMapTools.chebyshev_companion_matrix — Function

chebyshev_companion_matrix(v::AbstractVector)

Input:

v: The coefficients of a Chebyshev polynomial

Output:

C: The Chebyshev companion (colleague) matrix of polynomial v

SymplecticMapTools.cheb_roots_to_roots — Function

cheb_roots_to_roots(μs)

Input:

μs: The roots of the polynomial satisfying q((z+inv(z))/2) = p(z), where p is a palindromic polynomial

Output:

λs: The roots of p